backend/requirements.txt
flask
requests
numpy
faiss-cpu
pandas
sentence-transformers
scikit-learn
